+++ locale/lang-eo.js
@@ -290,7 +290,7 @@ SnapTranslator.dict.eo = {
     "Plain prototype labels": "Krudaj prototipaj etikedoj",
     "Play": "Aŭdigi",
     "Play sound": "Aŭdigi sonon",
-    "Please make sure your web browser is up to date and your camera is properly configured. Some browsers also require you to access Snap! through HTTPS to use the camera. Plase replace the \"http://\" part of the address in your browser by \"https://\" and try again.": "",
+    "Please make sure your web browser is up to date and your camera is properly configured. Some browsers also require you to access Snap! through HTTPS to use the camera. Please replace the \"http://\" part of the address in your browser by \"https://\" and try again.": "",
     "Please use the verification link that was sent to your email address when you signed up.": "",
     "Polygon": "",
     "Predicate": "Predikato",