+++ src/api.js
@@ -7,7 +7,7 @@
     written by Jens Mönig
     jens@moenig.org
 
-    Copyright (C) 2023 by Jens Mönig
+    Copyright (C) 2024 by Jens Mönig
 
     This file is part of Snap!.
 
@@ -39,14 +39,14 @@
 
 */
 
-/*global modules, IDE_Morph, isString, Map, List, world, isNil, Project,
-detect, isSnapObject, VariableFrame*/
+/*global modules, IDE_Morph, isString, Map, List, Project, detect, isSnapObject,
+VariableFrame*/
 
 /*jshint esversion: 11*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.api = '2023-January-30';
+modules.api = '2024-February-22';
 
 // IDE_Morph external communication API
 /*
@@ -54,20 +54,6 @@ modules.api = '2023-January-30';
     add message listeners to Snap! broadcasts and access
     global variables
 */
-
-window.onmessage = function (event) {
-    // make the API accessible from outside an iframe
-    var ide = world.children[0];
-    if (!isNil(event.data.selector)) {
-        window.top.postMessage(
-            {
-                selector: event.data.selector,
-                response: ide[event.data.selector].apply(ide, event.data.params)
-            },
-            '*'
-        );
-    }
-};
 
 IDE_Morph.prototype.getScenes = function () {
     // return an array of all scenenames
@@ -278,14 +264,31 @@ IDE_Morph.prototype.loadSpriteScriptsXML = function (scriptsXML) {
     return this.spriteNamed(name).synchScriptsFrom(scriptsXML);
 };
 
-IDE_Morph.prototype.flashSpriteScripts = function (fromLOC, toLOC, name) {
+IDE_Morph.prototype.flashSpriteScripts = function (fromLOC, toLOC, name, clr) {
     // highlight the blocks of the scripts of the sprite indicated by name or
     // the current sprite or stage if none that correspond to the portion of the
     // text between the start- and end lines when using the current codification
-    // mapping
+    // mapping.
+    // Optionally a string of comma-separated "r,g,b[,a]" values can be passed
+    // in to specify a specific highlight color, where each color component is
+    // a number between 0 and 255 and alpha is a fraction between 0 and 1.
+    // If none is supplied the default flash color is used.
     var scripts = this.spriteNamed(name).scripts;
     scripts.unflash();
-    scripts.flashLOC(fromLOC, toLOC);
+    scripts.flashLOC(fromLOC, toLOC, clr);
+};
+
+IDE_Morph.prototype.flashSpriteScriptAt = function (charIdx, name, clr) {
+    // highlight the innermost block of the scripts of the sprite indicated by
+    // name or the current sprite or stage if none that corresponds to the index
+    // of the text given the current codification mapping.
+    // Optionally a string of comma-separated "r,g,b[,a]" values can be passed
+    // in to specify a specific highlight color, where each color component is
+    // a number between 0 and 255 and alpha is a fraction between 0 and 1.
+    // If none is supplied the default flash color is used.
+    var scripts = this.spriteNamed(name).scripts;
+    scripts.unflash();
+    scripts.flashCodeIdx(charIdx, clr);
 };
 
 IDE_Morph.prototype.unflashSpriteScripts = function (name) {
@@ -294,8 +297,54 @@ IDE_Morph.prototype.unflashSpriteScripts = function (name) {
     this.spriteNamed(name).scripts.unflash();
 };
 
+IDE_Morph.prototype.flashSpriteScriptOutlineAt = function (
+    charIdx,
+    name,
+    clr,
+    border
+) {
+    // highlight the outline of the innermost block of the scripts of the sprite
+    // indicated by name or the current sprite or stage if none that corresponds
+    // to the index of the text given the current codification mapping.
+    // Optionally a string of comma-separated "r,g,b[,a]" values can be passed
+    // in to specify a specific highlight color, where each color component is
+    // a number between 0 and 255 and alpha is a fraction between 0 and 1.
+    // If none is supplied the default flash color is used.
+    // Also optionally a border width of pixels can be specified
+    var scripts = this.spriteNamed(name).scripts;
+    // scripts.unflash();
+    scripts.flashOutlineCodeIdx(charIdx, clr, border);
+};
+
+IDE_Morph.prototype.unflashSpriteScriptsOutline = function (name) {
+    // un-highlight the script outlines of the sprite indicated by name or the
+    // current sprite or stage if none
+    this.spriteNamed(name).scripts.unflashOutline();
+};
+
+IDE_Morph.prototype.showScriptBalloonAt = function (contents, charIdx, name) {
+    // popup a balloon at the innermost block of the scripts of the sprite
+    // indicated by name or the current sprite or stage if none that corresponds
+    // to the index of the text given the current codification mapping, and
+    // display the given contents, which can be a string, number, costume,
+    // morph, canvas, list, table etc. (anything first-class in Snap!)
+    var scripts = this.spriteNamed(name).scripts;
+    // scripts.unflash();
+    // scripts.flashCodeIdx(charIdx, contents);
+    scripts.balloonCodeIdx(charIdx, contents);
+};
+
+IDE_Morph.prototype.closePopUps = function () {
+    // remove all pop-up menus and balloons, if any
+    this.world().hand.destroyTemporaries();
+};
+
 IDE_Morph.prototype.unsavedChanges = function () {
     return this.hasUnsavedEdits();
+};
+
+IDE_Morph.prototype.resetUnsavedChanges = function () {
+    return this.recordSavedChanges();
 };
 
 IDE_Morph.prototype.setTranslation = function (countryCode, callback) {